+++ tests/test_end_to_end.py
@@ -79,6 +79,27 @@ class TestEndToEnd(unittest.TestCase):
         import os
         os.remove('test.pt')
 
+    def test_with_loader(self):
+        p = torch.tensor([2.0, 1.0, 10.0])
+        training_steps = 2
+
+        model = Net(p)
+        optim = torch.optim.SGD(model.parameters(), lr=0.01)
+        test_var = {'loaded': False}
+
+        def custom_loader(state):
+            state[torchbearer.X], state[torchbearer.Y_TRUE] = None, None
+            test_var['loaded'] = True
+
+        tbmodel = torchbearer.Trial(model, optim, loss, callbacks=[torchbearer.callbacks.MostRecent(filepath='test.pt')]).for_train_steps(training_steps).for_val_steps(1)
+        tbmodel.with_loader(custom_loader)
+        self.assertTrue(not test_var['loaded'])
+        tbmodel.run(1)
+        self.assertTrue(test_var['loaded'])
+
+        import os
+        os.remove('test.pt')
+
     def test_only_model(self):
         p = torch.tensor([2.0, 1.0, 10.0])