+++ tests/test_trial.py
@@ -658,13 +658,20 @@ class TestWithData(TestCase):
         td.assert_called_once_with(x)
 
 
-class TestWithClosure(TestCase):
+class TestWithClosureAndLoader(TestCase):
     def test_with_closure(self):
         def closure():
             return 'test'
         t = Trial(None)
         t.with_closure(closure)
         self.assertTrue(t.closure() == 'test')
+
+    def test_with_loader(self):
+        def loader(state):
+            print('test')
+        t = Trial(None)
+        t.with_loader(loader)
+        self.assertTrue(t.state[torchbearer.LOADER] == loader)
 
 
 class TestRun(TestCase):
@@ -900,7 +907,8 @@ class TestRun(TestCase):
 
 
 class TestFitPass(TestCase):
-    def test_fit_train_called(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_train_called(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])), (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
         train_steps = len(data)
@@ -913,7 +921,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion, torchbearer.OPTIMIZER: optimizer,
@@ -925,12 +933,13 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertEqual(torchbearertrial.train.call_count, 1)
 
-    def test_fit_metrics_reset(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_metrics_reset(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])), (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
         train_steps = len(data)
@@ -943,7 +952,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion, torchbearer.OPTIMIZER: optimizer,
@@ -955,12 +964,13 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertEqual(metric_list.reset.call_count, 1)
 
-    def test_fit_callback_calls(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_callback_calls(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])), (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
         train_steps = len(data)
@@ -974,7 +984,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion, torchbearer.OPTIMIZER: optimizer,
@@ -986,7 +996,7 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertEqual(callback_list.on_start_training.call_count, 1)
@@ -997,7 +1007,8 @@ class TestFitPass(TestCase):
         self.assertTrue(callback_list.on_step_training.call_count == 3)
         self.assertEqual(callback_list.on_end_training.call_count, 1)
 
-    def test_fit_optimizer_calls(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_optimizer_calls(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])), (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
         train_steps = len(data)
@@ -1011,7 +1022,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion, torchbearer.OPTIMIZER: optimizer,
@@ -1023,13 +1034,14 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(optimizer.zero_grad.call_count == 3)
         self.assertTrue(optimizer.step.call_count == 3)
 
-    def test_fit_forward_call_no_state(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_forward_call_no_state(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])), (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
         train_steps = len(data)
@@ -1043,7 +1055,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion, torchbearer.OPTIMIZER: optimizer,
@@ -1056,13 +1068,14 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = False
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(torchmodel.call_count == 3)
         self.assertTrue(torchmodel.call_args_list[0][0][0].item() == 1)
 
-    def test_fit_forward_call_with_state(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_forward_call_with_state(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])),
                 (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
@@ -1077,7 +1090,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
@@ -1092,13 +1105,14 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(torchmodel.call_count == 3)
         self.assertTrue(len(torchmodel.call_args_list[0][1]) == 1)
 
-    def test_fit_criterion(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_criterion(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])),
                 (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
@@ -1116,7 +1130,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
@@ -1124,15 +1138,15 @@ class TestFitPass(TestCase):
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float,
             torchbearer.HISTORY: [], torchbearer.TRAIN_GENERATOR: generator, torchbearer.TRAIN_STEPS: train_steps, torchbearer.EPOCH: 0,
-            torchbearer.BACKWARD_ARGS: {}
+            torchbearer.BACKWARD_ARGS: {}, torchbearer.GENERATOR: generator
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
-
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None,
+                                  torchbearer.GENERATOR: generator}
         torchbearertrial._fit_pass(state)
         self.assertTrue(criterion.call_count == 3)
         self.assertTrue(criterion.call_args_list[0][0][0] == 5)
@@ -1160,7 +1174,7 @@ class TestFitPass(TestCase):
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
-            torchbearer.OPTIMIZER: optimizer, torchbearer.INF_TRAIN_LOADING: False,
+            torchbearer.OPTIMIZER: optimizer, torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float,
             torchbearer.HISTORY: [], torchbearer.TRAIN_GENERATOR: generator, torchbearer.TRAIN_STEPS: train_steps, torchbearer.EPOCH: 0,
@@ -1170,14 +1184,16 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
-        torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
+        torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list, torchbearer.LOADER: None,
                                   torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(criterion.call_count == 3)
         self.assertTrue(criterion.call_args_list[0][0][0] == state)
 
-    def test_fit_backward(self):
+
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_backward(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])),
                 (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
@@ -1193,13 +1209,13 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
             torchbearer.OPTIMIZER: optimizer, torchbearer.INF_TRAIN_LOADING: False,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
-            torchbearer.DATA_TYPE: torch.float,
+            torchbearer.DATA_TYPE: torch.float, torchbearer.LOADER: None,
             torchbearer.HISTORY: [], torchbearer.TRAIN_GENERATOR: generator, torchbearer.TRAIN_STEPS: train_steps, torchbearer.EPOCH: 0,
             torchbearer.BACKWARD_ARGS: {}
         ]
@@ -1208,12 +1224,13 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.INF_TRAIN_LOADING: False, torchbearer.TRAIN_DATA: (generator, train_steps)}
+                                  torchbearer.INF_TRAIN_LOADING: False, torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(loss.backward.call_count == 3)
 
-    def test_fit_metrics_process(self):
+    @patch('torchbearer.CallbackListInjection')
+    def test_fit_metrics_process(self, mock_inj):
         data = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2])),
                 (torch.Tensor([3]), torch.Tensor([3]))]
         generator = DataLoader(data)
@@ -1228,7 +1245,7 @@ class TestFitPass(TestCase):
 
         metric_list = MagicMock()
         callback_list = MagicMock()
-        torchbearer.CallbackListInjection = Mock(return_value=callback_list)
+        mock_inj.return_value = callback_list
 
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
@@ -1242,7 +1259,7 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertTrue(metric_list.process.call_count == 3)
@@ -1278,7 +1295,7 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         history = torchbearertrial._fit_pass(state)[torchbearer.METRICS]
         self.assertEqual(metric_list.process_final.call_count, 1)
@@ -1315,7 +1332,7 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = True
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, train_steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         torchbearertrial._fit_pass(state)
         self.assertEqual(metric_list.process.call_count, 1)
@@ -1350,7 +1367,7 @@ class TestFitPass(TestCase):
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = False
-        torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: None, torchbearer.CALLBACK_LIST: callback_list, torchbearer.TRAIN_DATA: (None, steps)}
+        torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: None, torchbearer.CALLBACK_LIST: callback_list, torchbearer.TRAIN_DATA: (None, steps), torchbearer.LOADER: None}
 
         state = torchbearertrial._fit_pass(state)
         self.assertTrue(state[torchbearer.ITERATOR] is None)
@@ -1388,7 +1405,7 @@ class TestFitPass(TestCase):
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = False
         torchbearertrial.state = {torchbearer.TRAIN_GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list,
-                                  torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: False,}
+                                  torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: False, torchbearer.LOADER: None}
 
         state = torchbearertrial._fit_pass(state)
         self.assertTrue(state[torchbearer.ITERATOR] is not None)
@@ -1422,7 +1439,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1457,7 +1474,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1497,7 +1514,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_standard)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_standard
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1533,7 +1550,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1573,7 +1590,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_standard)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_standard
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1612,15 +1629,15 @@ class TestTestPass(TestCase):
         state = make_state[
             torchbearer.MAX_EPOCHS: epochs, torchbearer.STOP_TRAINING: False, torchbearer.MODEL: torchmodel, torchbearer.CRITERION: criterion,
             torchbearer.OPTIMIZER: optimizer,
-            torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
+            torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu', torchbearer.LOADER: None,
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_standard)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_standard
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
         torchbearertrial.train = Mock()
         torchbearertrial.pass_state = False
-        torchbearertrial.state = {torchbearer.GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list}
+        torchbearertrial.state = {torchbearer.GENERATOR: generator, torchbearer.CALLBACK_LIST: callback_list, torchbearer.LOADER: None,}
 
         torchbearertrial._test_pass(state)
         self.assertTrue(criterion.call_count == 3)
@@ -1650,7 +1667,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1685,7 +1702,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1721,7 +1738,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1756,7 +1773,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: None, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1791,7 +1808,7 @@ class TestTestPass(TestCase):
             torchbearer.OPTIMIZER: optimizer,
             torchbearer.METRIC_LIST: metric_list, torchbearer.CALLBACK_LIST: callback_list, torchbearer.DEVICE: 'cpu',
             torchbearer.DATA_TYPE: torch.float, torchbearer.HISTORY: [], torchbearer.GENERATOR: generator, torchbearer.STEPS: steps, torchbearer.EPOCH: 0,
-            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: torchbearer.trial.Sampler(load_batch_none)
+            torchbearer.X: data[0][0], torchbearer.Y_TRUE: data[0][1], torchbearer.SAMPLER: load_batch_none
         ]
 
         torchbearertrial = Trial(torchmodel, optimizer, criterion, [], callbacks=[])
@@ -1816,7 +1833,8 @@ class TestTrialValEvalPred(TestCase):
         t = Trial(MagicMock())
         eval_mock = t.eval = Mock()
         test_pass_mock = t._test_pass = Mock()
-        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None, torchbearer.VALIDATION_DATA: (generator, steps)}
+        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None,
+                   torchbearer.VALIDATION_DATA: (generator, steps), torchbearer.LOADER: None}
         metrics = t._validation_pass(state)
 
         self.assertEqual(eval_mock.call_count, 1)
@@ -1835,11 +1853,12 @@ class TestTrialValEvalPred(TestCase):
         t = Trial(MagicMock())
         eval_mock = t.eval = Mock()
         t._test_pass = Mock()
-        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None, torchbearer.VALIDATION_DATA: (generator, steps)}
+        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None,
+                   torchbearer.VALIDATION_DATA: (generator, steps), torchbearer.LOADER: None}
         t._validation_pass(state)
 
         self.assertTrue(eval_mock.call_count == 0)
-        
+
     def test_evaluate(self):
         generator = MagicMock()
         steps = 5
@@ -1848,7 +1867,9 @@ class TestTrialValEvalPred(TestCase):
         t = Trial(MagicMock())
         eval_mock = t.eval = Mock()
         clist = MagicMock()
-        state = {torchbearer.HISTORY: [('steps', {'train_metric': 2})], torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: clist, torchbearer.VALIDATION_STEPS: steps, torchbearer.VALIDATION_DATA: (generator, steps), torchbearer.METRICS: {'val_metric': 1}}
+        state = {torchbearer.HISTORY: [('steps', {'train_metric': 2})], torchbearer.VALIDATION_GENERATOR: generator,
+                 torchbearer.CALLBACK_LIST: clist, torchbearer.VALIDATION_STEPS: steps, torchbearer.VALIDATION_DATA: (generator, steps),
+                 torchbearer.METRICS: {'val_metric': 1}, torchbearer.LOADER: None}
         test_pass_mock = t._test_pass = Mock(return_value=state)
         t.state = state
         metrics = t.evaluate()
@@ -1873,7 +1894,8 @@ class TestTrialValEvalPred(TestCase):
         t = Trial(MagicMock())
         eval_mock = t.eval = Mock()
         test_pass_mock = t._test_pass = Mock(return_value={torchbearer.METRICS: 1})
-        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None, torchbearer.VALIDATION_STEPS: steps, torchbearer.VALIDATION_DATA: (generator, steps)}
+        t.state = {torchbearer.VALIDATION_GENERATOR: generator, torchbearer.CALLBACK_LIST: None,
+                   torchbearer.VALIDATION_STEPS: steps, torchbearer.VALIDATION_DATA: (generator, steps), torchbearer.LOADER: None}
         metrics = t.evaluate()
 
         self.assertTrue(eval_mock.call_count == 0)
@@ -1888,7 +1910,8 @@ class TestTrialValEvalPred(TestCase):
         eval_mock = t.eval = Mock()
         test_pass_mock = t._test_pass = Mock(return_value={torchbearer.FINAL_PREDICTIONS: 1})
         clist = MagicMock()
-        t.state = {torchbearer.TEST_GENERATOR: generator, torchbearer.CALLBACK_LIST: clist, torchbearer.TEST_STEPS: steps, torchbearer.TEST_DATA: (generator, steps)}
+        t.state = {torchbearer.TEST_GENERATOR: generator, torchbearer.CALLBACK_LIST: clist, torchbearer.TEST_STEPS: steps,
+                   torchbearer.TEST_DATA: (generator, steps), torchbearer.LOADER: None}
         metrics = t.predict(state)
 
         self.assertEqual(clist.on_start.call_count, 1)
@@ -1911,7 +1934,8 @@ class TestTrialValEvalPred(TestCase):
         t = Trial(MagicMock())
         eval_mock = t.eval = Mock()
         test_pass_mock = t._test_pass = Mock(return_value={torchbearer.FINAL_PREDICTIONS: 1})
-        t.state = {torchbearer.TEST_GENERATOR: generator, torchbearer.CALLBACK_LIST: None, torchbearer.TEST_STEPS: steps, torchbearer.TEST_DATA: (generator, steps)}
+        t.state = {torchbearer.TEST_GENERATOR: generator, torchbearer.CALLBACK_LIST: None, torchbearer.TEST_STEPS: steps,
+                   torchbearer.TEST_DATA: (generator, steps), torchbearer.LOADER: None}
         metrics = t.predict(state)
 
         self.assertTrue(eval_mock.call_count == 0)
@@ -2147,7 +2171,7 @@ class TestTrialMembers(TestCase):
         self.assertTrue(torchmodel.to.call_args[1]['dtype'] == dtype)
         self.assertTrue(state_tensor.to.call_args[1]['device'] == dev)
         self.assertTrue(state_tensor.to.call_args[1]['dtype'] == dtype)
-        
+
     @patch('torch.cuda.current_device')
     def test_cuda_no_device(self, device_mock):
         device_mock.return_value = 111
@@ -2189,7 +2213,7 @@ class TestTrialMembers(TestCase):
         torchbearertrial.cpu()
 
         self.assertTrue(torchbearertrial.to.call_args[0][0] == 'cpu')
-        
+
     def test_load_state_dict_resume(self):
         key_words = {'strict': True}
 
@@ -2441,42 +2465,59 @@ class TestTrialFunctions(TestCase):
         steps = None
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.GENERATOR)
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_standard)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.GENERATOR: (generator, steps)}
+        t.state = {torchbearer.GENERATOR: (generator, steps), torchbearer.LOADER: None}
         t.test_func()
-        self.assertTrue(t.state[torchbearer.SAMPLER].batch_loader == torchbearer.trial.load_batch_standard)
+        self.assertTrue(t.state[torchbearer.SAMPLER] == torchbearer.trial.load_batch_standard)
 
     def test_inject_sampler_none(self):
         generator = None
         steps = None
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.GENERATOR)
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_standard)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.GENERATOR: (generator, steps)}
+        t.state = {torchbearer.GENERATOR: (generator, steps), torchbearer.LOADER: None}
         t.test_func()
-        self.assertTrue(t.state[torchbearer.SAMPLER].batch_loader == torchbearer.trial.load_batch_none)
+        self.assertTrue(t.state[torchbearer.SAMPLER] == torchbearer.trial.load_batch_none)
 
     def test_inject_sampler_predict(self):
         generator = MagicMock()
         steps = None
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.GENERATOR, predict=True)
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_predict)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.GENERATOR: (generator, steps)}
+        t.state = {torchbearer.GENERATOR: (generator, steps), torchbearer.LOADER: None}
         t.test_func()
-        self.assertTrue(t.state[torchbearer.SAMPLER].batch_loader == torchbearer.trial.load_batch_predict)
+        self.assertTrue(t.state[torchbearer.SAMPLER] == torchbearer.trial.load_batch_predict)
+
+    def test_inject_sampler_custom(self):
+        generator = MagicMock()
+        steps = None
+
+        class SomeClass:
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_predict)
+            def test_func(self):
+                pass
+
+        def some_loader(state):
+            return 'test'
+
+        t = SomeClass()
+        t.state = {torchbearer.GENERATOR: (generator, steps), torchbearer.LOADER: some_loader}
+        t.test_func()
+        self.assertTrue(t.state[torchbearer.SAMPLER] == some_loader)
 
     @patch('warnings.warn')
     @patch('torchbearer.trial.load_batch_infinite')
@@ -2485,12 +2526,12 @@ class TestTrialFunctions(TestCase):
         steps = -1
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.GENERATOR, predict=True)
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_predict)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.GENERATOR: (generator, steps)}
+        t.state = {torchbearer.GENERATOR: (generator, steps), torchbearer.LOADER: None}
         t.test_func()
         self.assertTrue(mock_lbi.call_args[0][0] == load_batch_predict)
 
@@ -2510,15 +2551,15 @@ class TestTrialFunctions(TestCase):
         steps = 10
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.TRAIN_DATA, predict=False)
+            @torchbearer.inject_sampler(torchbearer.TRAIN_DATA, load_batch_standard)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: True}
+        t.state = {torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: True, torchbearer.LOADER: None}
         t.test_func()
         self.assertTrue(mock_lbi.call_args[0][0] == load_batch_standard)
-        self.assertTrue(generator.inf)
+        self.assertTrue(generator.tb_iter)
 
     @patch('torchbearer.trial.load_batch_infinite')
     def test_inject_sampler_infinite_train_loading(self, mock_lbi):
@@ -2527,12 +2568,12 @@ class TestTrialFunctions(TestCase):
         steps = 5
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.TRAIN_DATA, predict=False)
+            @torchbearer.inject_sampler(torchbearer.TRAIN_DATA, load_batch_standard)
             def test_func(self):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: True}
+        t.state = {torchbearer.TRAIN_DATA: (generator, steps), torchbearer.INF_TRAIN_LOADING: True, torchbearer.LOADER: None}
         t.test_func()
         self.assertTrue(mock_lbi.call_args[0][0] == load_batch_standard)
 
@@ -2542,12 +2583,12 @@ class TestTrialFunctions(TestCase):
         test_steps = 1
 
         class SomeClass:
-            @torchbearer.inject_sampler(torchbearer.GENERATOR, predict=False)
+            @torchbearer.inject_sampler(torchbearer.GENERATOR, load_batch_predict)
             def test_func(self, data_key=None):
                 pass
 
         t = SomeClass()
-        t.state = {torchbearer.GENERATOR: (generator, None), torchbearer.TEST_GENERATOR: (test_generator, test_steps)}
+        t.state = {torchbearer.GENERATOR: (generator, None), torchbearer.TEST_GENERATOR: (test_generator, test_steps), torchbearer.LOADER: None}
         t.test_func(data_key=torchbearer.TEST_GENERATOR)
         self.assertTrue(t.state[torchbearer.GENERATOR] == test_generator)
         self.assertTrue(t.state[torchbearer.STEPS] == test_steps)
@@ -2568,7 +2609,8 @@ class TestTrialFunctions(TestCase):
         self.assertTrue(c_inj.call_args[0][0] == test_callback)
 
     def test_deep_to_tensor(self):
-        tensor = MagicMock()
+        base_tensor = torch.Tensor([1])
+        tensor = MagicMock(spec=base_tensor)
         new_dtype = torch.float16
         new_device = 'cuda:1'
 
@@ -2577,7 +2619,8 @@ class TestTrialFunctions(TestCase):
         self.assertTrue(tensor.to.call_args[0][1] == new_dtype)
 
     def test_deep_to_tensor_int_dtype(self):
-        tensor = MagicMock()
+        base_tensor = torch.Tensor([1])
+        tensor = MagicMock(spec=base_tensor)
         tensor.dtype = torch.uint8
         new_device = 'cuda:1'
         new_dtype = torch.uint8
@@ -2587,8 +2630,9 @@ class TestTrialFunctions(TestCase):
         self.assertTrue(len(tensor.to.call_args[0]) == 1)
 
     def test_deep_to_list(self):
-        tensor_1 = MagicMock()
-        tensor_2 = MagicMock()
+        base_tensor = torch.Tensor([1])
+        tensor_1 = MagicMock(spec=base_tensor)
+        tensor_2 = MagicMock(spec=base_tensor)
         tensors = [tensor_1, tensor_2]
         new_dtype = torch.float16
         new_device = 'cuda:1'
@@ -2612,6 +2656,17 @@ class TestTrialFunctions(TestCase):
         self.assertTrue(tensor_1.to.call_args[0][1] == new_dtype)
         self.assertTrue(tensor_2.to.call_args[0][0] == new_device)
         self.assertTrue(tensor_2.to.call_args[0][1] == new_dtype)
+
+    def test_deep_to_unknown_object(self):
+        tensor_1 = MagicMock()
+        tensor_2 = MagicMock()
+        tensors = {'t1': tensor_1, 't2': tensor_2}
+        new_dtype = torch.float16
+        new_device = 'cuda:1'
+
+        deep_to(tensors, new_device, new_dtype)
+        self.assertTrue(tensor_1.to.call_args is None)
+        self.assertTrue(tensor_2.to.call_args is None)
 
     def test_load_batch_standard(self):
         items = [(torch.Tensor([1]), torch.Tensor([1])), (torch.Tensor([2]), torch.Tensor([2]))]