+++ torchbearer/trial.py
@@ -163,7 +163,7 @@ def deep_to(batch, device, dtype):
     elif isinstance(batch, dict):
         for key in batch:
             batch[key] = deep_to(batch[key], device, dtype)
-    else:
+    elif torch.is_tensor(batch):
         if batch.dtype.is_floating_point:
             batch = batch.to(device, dtype)
         else:
@@ -222,59 +222,52 @@ def load_batch_predict(state):
         state[torchbearer.X] = data
 
 
-class Sampler:
-    """
-    Sampler wraps a batch loader function and executes it when :meth:`Sampler.sample` is called
-
-    Args:
-        batch_loader (func): The batch loader to execute
-    """
-    def __init__(self, batch_loader):
-        self.batch_loader = batch_loader
-
-    def sample(self, state):
-        self.batch_loader(state)
-
-
-def inject_sampler(data_key, predict=False):
+def inject_sampler(data_key, batch_sampler):
     """ Decorator to inject a :class:`Sampler` into state[torchbearer.SAMPLER] along with the specified \
         generator into state[torchbearer.GENERATOR] and number of steps into state[torchbearer.STEPS]
 
     Args:
-        data_key (StateKey): :class:`.StateKey` for the data to inject
-        predict (bool): If true, the prediction batch loader is used, if false the standard data loader is used
+        data_key (:class:`.StateKey`): Key for the data to inject
+        batch_sampler (function): Batch sampler function that extracts batch from data loader, stores in state and sends
+        data to correct device
 
     Returns:
         The decorator
     """
     def decorator(func):
-        @functools.wraps(func)
-        def wrapper(self, *args, **kwargs):
-            key = kwargs['data_key'] if 'data_key' in kwargs else data_key  # Populate default value
-            generator, steps = self.state[key] if self.state[key] is not None else (None, None)
-
-            if generator is None:
-                loader = load_batch_none
-            elif predict:
-                loader = load_batch_predict
-            else:
-                loader = load_batch_standard
-
+        def infinite_wrapper(self, key, generator, steps, sampler):
             if generator is not None and steps is not None:
                 over_steps = steps > len(generator)
                 inf_steps = steps == -1
                 inf_train_loader = key == torchbearer.TRAIN_DATA and self.state[torchbearer.INF_TRAIN_LOADING]
-                if over_steps or inf_steps or inf_train_loader:  # Want iterator to refresh at end
+
+                if over_steps or inf_steps or inf_train_loader:  # Want iterator to refresh at end not per epoch
                     if steps == -1: warnings.warn("Trial is set to run indefinitely. "
                                               "Make sure you have some method to terminate safely.")
-                    loader = load_batch_infinite(loader)
+                    sampler = load_batch_infinite(sampler)
 
-                if inf_train_loader and not hasattr(generator, 'inf'):  # First run and want iterator to not refresh each epoch but on end
-                    generator.inf = True
+                # Want iterator to run until end before refreshing regardless of number of train/val steps
+                if inf_train_loader and not hasattr(generator, 'tb_iter'):
                     generator.tb_iter = iter(generator)
 
+            return generator, sampler
+
+        @functools.wraps(func)
+        def wrapper(self, *args, **kwargs):
+            sampler = batch_sampler
+
+            key = kwargs['data_key'] if 'data_key' in kwargs else data_key  # Populate default value
+            generator, steps = self.state[key] if self.state[key] is not None else (None, None)
+
+            if self.state[torchbearer.LOADER] is not None:
+                sampler = self.state[torchbearer.LOADER]
+            elif generator is None:
+                sampler = load_batch_none
+
+            generator, sampler = infinite_wrapper(self, key, generator, steps, sampler)
+
             self.state[torchbearer.DATA] = key
-            self.state[torchbearer.SAMPLER] = Sampler(loader)
+            self.state[torchbearer.SAMPLER] = sampler
             self.state[torchbearer.GENERATOR] = generator
             self.state[torchbearer.STEPS] = steps
 
@@ -380,6 +373,7 @@ class Trial(object):
             torchbearer.VALIDATION_DATA: None,
             torchbearer.TEST_DATA: None,
             torchbearer.INF_TRAIN_LOADING: False,
+            torchbearer.LOADER: None
         })
 
         self.state[torchbearer.CALLBACK_LIST].on_init(self.state)
@@ -677,11 +671,25 @@ class Trial(object):
 
         return self
 
+    def with_loader(self, batch_loader):
+        """Use this trial with custom batch loader. Usually calls next on state[torchbearer.ITERATOR] and populates
+        state[torchbearer.X] and state[torchbearer.Y_TRUE]
+
+        Args:
+            batch_loader (function): Function of state that extracts data from data loader (stored under torchbearer.ITERATOR), stores it in
+            state and sends it to the correct device
+
+        Returns:
+            Trial: self:
+        """
+        self.state[torchbearer.LOADER] = batch_loader
+        return self
+
     def with_closure(self, closure):
         """Use this trial with custom closure
-        
+
         Args:
-            closure (function): Function of state that defines the custom closure 
+            closure (function): Function of state that defines the custom closure
 
         Returns:
             Trial: self:
@@ -755,7 +763,7 @@ class Trial(object):
         else:
             return iter(generator)
 
-    @inject_sampler(torchbearer.TRAIN_DATA)
+    @inject_sampler(torchbearer.TRAIN_DATA, load_batch_standard)
     def _fit_pass(self, state):
         state.update(self.state)  # TODO: Hack to make injection work, should be removed if `self.state` is mutable
         self.train()
@@ -767,7 +775,7 @@ class Trial(object):
 
         state[torchbearer.CALLBACK_LIST].on_start_training(state)
         for state[torchbearer.BATCH] in (range(state[torchbearer.STEPS]) if state[torchbearer.STEPS] != -1 else itertools.count()):
-            state[torchbearer.SAMPLER].sample(state)
+            state[torchbearer.SAMPLER](state)
             state[torchbearer.CALLBACK_LIST].on_sample(state)
 
             # Update parameters
@@ -786,7 +794,7 @@ class Trial(object):
 
     def _test_pass(self, state):
         with torch.no_grad():
-            state[torchbearer.ITERATOR] = iter(state[torchbearer.GENERATOR]) if state[torchbearer.GENERATOR] is not None else None  # TODO: Inject this?
+            state[torchbearer.ITERATOR] = Trial._new_iter(state[torchbearer.GENERATOR])
 
             state[torchbearer.METRIC_LIST].reset(state)
             state[torchbearer.METRICS] = {}
@@ -794,7 +802,7 @@ class Trial(object):
             state[torchbearer.CALLBACK_LIST].on_start_validation(state)
 
             for state[torchbearer.BATCH] in range(state[torchbearer.STEPS]):
-                state[torchbearer.SAMPLER].sample(state)
+                state[torchbearer.SAMPLER](state)
                 state[torchbearer.CALLBACK_LIST].on_sample_validation(state)
 
                 # Forward Pass
@@ -825,7 +833,7 @@ class Trial(object):
             state[torchbearer.CALLBACK_LIST].on_end_validation(state)
         return state
 
-    @inject_sampler(torchbearer.VALIDATION_DATA)
+    @inject_sampler(torchbearer.VALIDATION_DATA, load_batch_standard)
     def _validation_pass(self, state):
         state.update(self.state)  # TODO: Hack to make injection work, should be removed if `self.state` is mutable
 
@@ -835,7 +843,7 @@ class Trial(object):
             self._test_pass(state)
         return state[torchbearer.METRICS]
 
-    @inject_sampler(torchbearer.VALIDATION_DATA)
+    @inject_sampler(torchbearer.VALIDATION_DATA, load_batch_standard)
     @inject_printer(validation_label_letter='e')
     def evaluate(self, verbose=-1, data_key=None):  # Note: kwargs appear unused but are inspected in inject_sampler
         """Evaluate this trial on the validation data.
@@ -872,7 +880,7 @@ class Trial(object):
         return {}
 
     @inject_callback(AggregatePredictions())
-    @inject_sampler(torchbearer.TEST_DATA, predict=True)
+    @inject_sampler(torchbearer.TEST_DATA, load_batch_predict)
     @inject_printer(validation_label_letter='p')
     def predict(self, verbose=-1, data_key=None):  # Note: kwargs appear unused but are inspected in inject_sampler
         """Determine predictions for this trial on the test data.